+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+ - Added an ICommercialRoom extensions class which has a method for determing the contextual volume point the room should be using.
 
 ## [14.1.1] - 2019-11-20
 ### Changed